Plot the results of "hot hand" simulations

Plot the results of "hot hand" simulations

Usage

plot_hothand_simulation(sim_results)

plot_hothand_trends(trend_results)

Arguments

sim_results

[numeric(n)] A vector with post-streak chomp-rates for multiple simulations with the same simulation parameters.

trend_results

[data.frame] A data frame with columns chomp_attempts, chomp_rate, and streak_rate. Each row represents the results from a single hot hand simulation. The simulation parameters for each row are described by columns chomp_attempts and chomp_rate.

Value

[ggplot] A ggplot of:

plot_hothand_trends(): the mean "post-streak" chomp rate (solid lines) over chomp attempts (dashed lines) for different independent chomp rates.

plot_hothand_simulation(): a histogram of "post-streak" chomp rates with the mean simulated rate represented by a solid red line and the independent chomp rate indicated by a dashed red line.