Kenai Fjords Temperature Readings

Temperature readings from five sites in Kenai Fjords 2007-2024.

Usage

kefj_temperature

kefj_site

kefj_datetime

kefj_tidelevel

kefj_exposure

kefj_season

Format

kefj_* are all vectors of length 2187966.

kefj_temperature A numeric vector with the temperature readings in Celsius.

kefj_site A character vector with the name of the site (one of Aialik, Harris, McCarty, Nuka_Bay, or Nuka_Pass) of the temperature reading.

kefj_datetime A POSIXct vector with the date and time of the temperature record.

kefj_tidelevel A numeric vector with the predicted tide level (meters) at the time of the temperature reading.

kefj_exposure A character vector indicating the type of temperature reading:

air

The predicted tidal elevation was at least 0.5 m lower than the tide height at which the logger was deployed

transition

The predicted tidal elevation was within 0.5 m of the tide height at which the logger was deployed

water

The predicted tidal elevation was at least 0.5 m higher than the tide height at which the logger was deployed

air/transition

The tidal elevation and tide height do not match the parameters for air or transition temperature and the temperature reading does not align with water temperature

kefj_season A five-level factor indicating season, relating to mussel life history:

Late winter

January–March, mussel gonad development

Spring

April–May, spawning and settlement

Summer

June–July, spawning and early growth

Fall

August–October, spawning and late growth

Early winter

November–December, senescence and early gonad development

Source

USGS_2024marinecs100b